Add logging and execution time aspect
A new aspect and annotation for logging and measuring the execution time of methods is introduced. This aspect is applied to various methods in the ProductsController, DeviceController, etc. The pom.xml was updated to include related dependencies, and logging level for the new aspect was added in the application YAML files. Also, the pom.xml of auth module was refactored.