Refactor indentation in Helm deployment template
The code refactor involved correcting the indentation of a line within the Helm deployment template. Specifically, this was done for the "instrumentation.opentelemetry.io/inject-java" annotation under the condition if ".Values.snomio.telemetry.enabled" is true. The change is purely cosmetic and does not affect the functionality of the deployment manifest.