Update default navigation to tasks page
Implemented a redirection in the routing configuration to default to the tasks page whenever an undefined route is accessed. The 'Navigate' component from 'react-router-dom' was imported and utilized to replace the previous placeholder message of 'Coming soon to a computer near you!'. User experience should now be improved by guiding users directly to the tasks page.