Rabbit entity refactor (#925)
Several code improvements were implemented. The DTO classes have been moved to their appropriate packages, improving code organization. Changes in some areas of the code improved the use of Lombok for simplification. A few test changes were performed to enhance reliability, introducing awaitility for condition-based waiting. This commit also adjusts the data type for more accurate comparisons in tests.

* Make product classes Serializable

Added Serializable interface to the 'ProductSummary', 'ProductPackSizes', 'BrandWithIdentifiers', 'ProductBrands' and 'ExternalIdentifier' classes. These changes allow storing instances of these classes in a serialized form or transferring them over a network and de-serializing them when necessary.

The code deletes the `MinimalTicketDto` class as it was deemed unnecessary. Various DTO classes in the `ticket` package have been updated with additional annotations. The major change is the addition of the `@ToString` annotation, which includes superclass fields in the output. Other adjustments include shuffling of existing annotations and the usage of `@SuperBuilder` in some case which overrides the regular `@Builder` usage.

* Add getComments API endpoint and mapping functions

A getComments endpoint has been added to the CommentController, allowing retrieval of all comments on a ticket. CommentMapper was updated to create a function that converts a list of Comment objects to a list of CommentDto objects. This change supports the serialization of Comment objects to CommentDto when fetching comments from the database and sending the response to the client.

* Add exception handling in MedicationNewBrandPackTest

An import for SingleConceptExpectedProblem exception has been added, and the product model retrieval process within the test now includes a try-catch block. This modification was made to handle SingleConceptExpectedProblem exceptions that may occur when retrieving the product model.

* Add build profile for arm64 architecture

A new Maven profile has been added to the pom.xml configuration file. This profile specifically targets building an image for the arm64 architecture when using Apple Silicon. It utilizes the `jib-maven-plugin` with `eclipse-temurin:17` as the base image and defines `arm64` as the architecture and `linux` as the operating system.

* Initialize sets and lists in Ticket model

The commit initializes the sets and lists declared in the Ticket model to prevent null pointer exceptions. It ensures these fields are never null by initializing them to empty instances of their corresponding types, i.e., HashSet or ArrayList. Adds @OrderBy annotation to maintain order by date created for comments and attachments.

* Refactor TicketService and improve resource handling

The commit includes several refactorings of the TicketService class. It cleans up and optimizes the handling of products, bulk product actions, and tickets. The changes involve the creation of helper methods for commonly performed operations, reducing repetition, and increasing code clarity. Additionally, potential issues of null handling were addressed, and redundant code was removed.

* Update return type in getTicket method

The method getTicket in TicketController now returns a TicketDto instead of a Ticket. This updated change will ensure better encapsulation and data transfer efficiency.

* Initialize conceptIds in BulkProductAction model

The Set<Long> variable conceptIds is now initialized to a new HashSet instance. This will prevent potential null pointer exceptions and avoid issues when trying to manipulate the set before initializing it.

* Add TicketAssociationService and enhance related classes

A new service class, TicketAssociationService, has been added to handle ticket association related operations. The TicketAssociationMapper has been updated with new methods and mappings. Also, the TicketAssociationController and

* Add UnknownAttributeTypeProblem and BaseAuditableDto classes

A new `UnknownAttributeTypeProblem` exception class was added in the snomio package to handle unknown attribute types. In the tickets package, a new abstract `BaseAuditableDto` class was introduced to hold common fields and behaviors for data transfer objects, such as id, version, creation and modification specifics.

* Implement second-level cache and enable query cache

Added Hibernate's second-level cache and query cache to improve performance. All necessary configurations have been added, including adding CacheConcurrencyStrategy.READ_WRITE to entities, enabling second-level cache and query cache in application.properties, and adding dependencies for ehcache in pom.xml. Additionally, added @JsonIgnore to getBrands() and getPackSizes() in BulkProductActionDto to ignore these methods during serialization/deserialization.

A new ehcache configuration file `ehcache.xml` has been added to the project. This configuration includes settings for two different types of caches and enables statistics and management features provided by JSR-107. In this configuration, the expiry time to live (TTL) is set to 120 seconds and heap size is set to 10,000 entries for both cache types.

* Update Prometheus metrics export configuration

The configuration key for enabling Prometheus metrics export has been updated. This change is critical to ensure that the application's metrics are correctly exported and used by Prometheus for monitoring.

* Add error suppression annotations and ticket associations

The primary change in this commit is the inclusion of error suppression annotations in multiple model classes to improve readability and maintain better adherence to coding conventions. Additionally, new 'ticketSourceAssociations' and 'ticketTargetAssociations' sets were added to the Ticket class. The existing ticket handling in the TicketServiceImpl class was also simplified. Refactorings were conducted to remove redundant or commented-out code.

* Refactor AdditionalFieldService and AdditionalFieldController

Refactoring was focused on AdditionalFieldService and AdditionalFieldController to improve code readability and structure. The functionality remains intact. Web-related elements were removed from the service class, which now only returns DTOs. Meanwhile, the controller class was simplified and now uses AdditionalFieldService methods. Hence, the controller is now purely responsible for HTTP request handling.

* added a patch method to /api/tickets/{ticketId}, primarily for our external system syncing, so they don't have to care so much about how the ticket currently looks, and instead they can care about just purely how the ticket should look from it's data, because that is the primary source of truth. This allows additional things that users have added to the tickets to now be removed.

* front end changes to start using the patch method where possible.

Co-authored-by: Attila Edelenyi <ede_ati@hotmail.com>

* format

---------

Co-authored-by: Dion McMurtrie <dion.mcmurtrie@csiro.au>
Co-authored-by: Edelenyi <attila.edelenyi@csiro.au>
Co-authored-by: Clinton Gillespie <clinton.gillespie@csiro.au>
Co-authored-by: dionmcm <dion.mcmurtrie@gmail.com>
Co-authored-by: Clinton Gillespie <clintongillespie@outlook.com>